SKLearn models refactored
