+++ bella/models/base.py
@@ -14,25 +14,13 @@ Mixin classes:
 from abc import ABC, abstractmethod
 from collections import defaultdict
 import copy
-from typing import Any, List, Callable, Dict, Union
+from typing import Any, List, Dict, Union
 from pathlib import Path
 
 import numpy as np
 from sklearn.externals import joblib
-from sklearn.pipeline import FeatureUnion, Pipeline
-from sklearn.preprocessing import MinMaxScaler
-from sklearn.svm import LinearSVC
 
 import bella
-from bella.tokenisers import ark_twokenize
-from bella.neural_pooling import matrix_max, matrix_min, matrix_avg
-from bella.neural_pooling import matrix_median, matrix_prod, matrix_std
-from bella.scikit_features.context import Context
-from bella.scikit_features.lexicon_filter import LexiconFilter
-from bella.scikit_features.join_context_vectors import JoinContextVectors
-from bella.scikit_features.neural_pooling import NeuralPooling
-from bella.scikit_features.tokeniser import ContextTokeniser
-from bella.scikit_features.word_vector import ContextWordVectors
 
 
 class BaseModel(ABC):
@@ -50,6 +38,7 @@ class BaseModel(ABC):
     2. predict -- Predict class labels for samples in X.
     3. probabilities -- The probability of each class label for all samples
        in X.
+    4. __repr__ -- Name of the machine learning model.
 
     Functions:
 
@@ -86,6 +75,15 @@ class BaseModel(ABC):
         :param X: Test samples matrix, shape = [n_samples, n_features]]
         :return: Probability of each class label for all samples, shape = \
         [n_samples, n_classes]
+        '''
+        pass
+
+    @abstractmethod
+    def __repr__(self) -> str:
+        '''
+        Name of the machine learning model.
+
+        :return: Name of the machine learning model.
         '''
         pass
 
@@ -191,23 +189,17 @@ class SKLearnModel(BaseModel):
     2. predict -- Predict class labels for samples in X.
     3. probabilities -- The probability of each class label for all samples
        in X.
+    4. __repr__ -- Name of the machine learning model.
 
     Functions:
 
     1. save -- Given a instance of this class will save it to a file.
     2. load -- Loads an instance of this class from a file.
-
-    Abstract Functions:
-
-    1. pipeline -- An instance :py:class:`sklearn.pipeline.Pipeline`
-       that represents the machine learning model of this class. Used in the
-       __init__ to set the model parameter.
     '''
 
     def __init__(self, *args, **kwargs) -> None:
 
         self.fitted = False
-        self.model = self.pipeline()
         self._model_parameters = self.get_parameters(*args, **kwargs)
         self.model.set_params(**self._model_parameters)
         self.fitted = False
@@ -273,20 +265,6 @@ class SKLearnModel(BaseModel):
         self._model_parameters = self.get_parameters(**value)
         self.model.set_params(**self._model_parameters)
         self.fitted = False
-
-    @staticmethod
-    @abstractmethod
-    def pipeline() -> Pipeline:
-        '''
-        An instance :py:class:`sklearn.pipeline.Pipeline` that represents the
-        machine learning model of this class. Used in the __init__ to set
-        the model parameter.
-
-        The pipeline instance may not be used in some cases stright away as
-        some of its transformers and estimator may not have parameters set.
-        E.g. which tokeniser to use.
-        '''
-        pass
 
     @staticmethod
     def save(model: 'bella.models.base.SKLearnModel',
@@ -455,794 +433,3 @@ class SKLearnModel(BaseModel):
         for param_dict in params_list:
             param_dict[param_name] = param_value
         return params_list
-
-
-class TargetInd(SKLearnModel):
-    '''
-    Target-ind model from `Vo and Zhang 2015 paper
-    <https://ijcai.org/Proceedings/15/Papers/194.pdf>`_.
-    '''
-
-    @classmethod
-    def get_cv_parameters(cls, word_vectors, tokeniser=None, lower=None,
-                          C=None, scale=None, random_state=None):
-        '''
-        Transform the given parameters into a list of dictonaries that is
-        accepted as `param_grid` parameter in
-        :py:class:`sklearn.model_selection.GridSearchCV`
-
-        :param word_vectors: A list of a list of word vectors e.g. [[SSWE()],
-                             [SSWE(), GloveCommonCrawl()]]
-        :param tokenisers: A list of tokeniser to be used e.g.
-                           :py:meth:`str.split`. Default None, only use default
-                           tokeniser.
-        :param lowers: A list of bool values which indicate whether to lower
-                       case the input words. Default None, only use default
-                       setting(True).
-        :param C: A list of C values for the :py:class:`sklearn.svm.SVC`
-                  estimator that is used in the pipeline. Default None, only
-                  use default setting(C=0.01).
-        :param random_state: A list of random_state values for the
-                             :py:class:`sklearn.svm.SVC` estimator that is
-                             used in the pipeline. Default None, only use
-                             default setting(random_state=42).
-        :param scale: List of scale values. The list can include
-                      :py:class:`sklearn.preprocessing.MinMaxScaler` type of
-                      clases or None if no scaling is to be used. Default None,
-                      only use default setting
-                      (:py:class:`sklearn.preprocessing.MinMaxScaler`).
-        :return: Parameters to explore through cross validation
-        '''
-        params_list = []
-        params_list = cls._add_to_params(params_list, word_vectors,
-                                         cls._get_word_vector_names())
-        if tokeniser is not None:
-            tokenisers_names = [param_name + '__tokeniser'
-                                for param_name in cls._get_tokeniser_names()]
-            params_list = cls._add_to_params(params_list, tokeniser,
-                                             tokenisers_names)
-        if lower is not None:
-            lower_names = [param_name + '__lower'
-                           for param_name in cls._get_tokeniser_names()]
-            params_list = cls._add_to_params(params_list, lower, lower_names)
-        if C is not None:
-            params_list = cls._add_to_all_params(params_list, 'svm__C', C)
-        if random_state is not None:
-            params_list = cls._add_to_all_params(params_list,
-                                                 'svm__random_state',
-                                                 random_state)
-        if scale is not None:
-            params_list = cls._add_to_all_params(params_list, 'scale', scale)
-        return params_list
-
-    @classmethod
-    def get_parameters(cls,
-                       word_vectors: List['bella.word_vectors.WordVectors'],
-                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
-                       lower: bool = True, C: float = 0.01,
-                       random_state: int = 42,
-                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
-        '''
-        Transform the given parameters into a dictonary that is accepted as
-        model parameters
-
-        :param word_vectors: A list of one or more word vectors to be used as
-                             feature vector lookups. If more than one is used
-                             the word vectors are concatenated together to
-                             create a the feature vector for each word.
-        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
-        :param lower: Whether to lower case the words
-        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
-                  that is used in the pipeline.
-        :param random_state: The random_state value for the
-                             :py:class:`sklearn.svm.SVC` estimator that is
-                             used in the pipeline.
-        :param scale: How to scale the data before input into the estimator.
-                      If no scaling is to be used set this to None.
-        :return: Model parameters
-        '''
-        params_dict = {}
-        # Add word vectors to Pipeline model
-        params_dict = cls._add_to_params_dict(params_dict,
-                                              cls._get_word_vector_names(),
-                                              word_vectors)
-        # Add tokenisers to Pipeline model
-        tokenisers_names = [param_name + '__tokeniser'
-                            for param_name in cls._get_tokeniser_names()]
-        params_dict = cls._add_to_params_dict(params_dict,
-                                              tokenisers_names,
-                                              tokeniser)
-        # Add if the words should be lower cased
-        lower_names = [param_name + '__lower'
-                       for param_name in cls._get_tokeniser_names()]
-        params_dict = cls._add_to_params_dict(params_dict, lower_names, lower)
-        # Add how the data should be scaled before going into the SVM
-        # If None then it means no scaling happens
-        params_dict = cls._add_to_params_dict(params_dict, ['scale'],
-                                              scale)
-        # Add the C value for the SVM
-        params_dict = cls._add_to_params_dict(params_dict, ['svm__C'], C)
-        # Add the random state for the SVM
-        params_dict = cls._add_to_params_dict(params_dict,
-                                              ['svm__random_state'],
-                                              random_state)
-        return params_dict
-
-    @staticmethod
-    def pipeline() -> Pipeline:
-        return Pipeline([
-            ('contexts', Context('full')),
-            ('tokens', ContextTokeniser()),
-            ('word_vectors', ContextWordVectors()),
-            ('pool_funcs', FeatureUnion([
-                ('max_pipe', Pipeline([
-                    ('max', NeuralPooling(matrix_max)),
-                    ('join', JoinContextVectors(matrix_median))
-                ])),
-                ('min_pipe', Pipeline([
-                    ('min', NeuralPooling(matrix_min)),
-                    ('join', JoinContextVectors(matrix_median))
-                ])),
-                ('avg_pipe', Pipeline([
-                    ('avg', NeuralPooling(matrix_avg)),
-                    ('join', JoinContextVectors(matrix_median))
-                ])),
-                ('prod_pipe', Pipeline([
-                    ('min', NeuralPooling(matrix_prod)),
-                    ('join', JoinContextVectors(matrix_median))
-                ])),
-                ('std_pipe', Pipeline([
-                    ('min', NeuralPooling(matrix_std)),
-                    ('join', JoinContextVectors(matrix_median))
-                ]))
-            ])),
-            ('scale', MinMaxScaler()),
-            ('svm', LinearSVC())
-        ])
-
-    @staticmethod
-    def _get_word_vector_names() -> List[str]:
-        '''
-        Method to be overidden by subclasses as each pipeline will be different
-        and will have different parameter names for where the word vectors are
-        set.
-
-        :returns: A list of of parameter names where the word vectors are set
-                  in the pipeline.
-        '''
-
-        return ['word_vectors__vectors']
-
-    @staticmethod
-    def _get_tokeniser_names() -> List[str]:
-        '''
-        Method to be overidden by subclasses as each pipeline will be different
-        and will have different parameter names for where tokenisers are
-        used.
-
-        :returns: A list of of parameter names where the tokenisers are used
-                  in the pipeline.
-        '''
-
-        return ['tokens']
-
-
-class TargetDepMinus(TargetInd):
-    '''
-    Target-dep- model from `Vo and Zhang 2015 paper
-    <https://ijcai.org/Proceedings/15/Papers/194.pdf>`_.
-    '''
-
-    @classmethod
-    def get_parameters(cls,
-                       word_vectors: List['bella.word_vectors.WordVectors'],
-                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
-                       lower: bool = True, C: float = 0.025,
-                       random_state: int = 42,
-                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
-        '''
-        Transform the given parameters into a dictonary that is accepted as
-        model parameters
-
-        :param word_vectors: A list of one or more word vectors to be used as
-                             feature vector lookups. If more than one is used
-                             the word vectors are concatenated together to
-                             create a the feature vector for each word.
-        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
-        :param lower: Wether to lower case the words
-        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
-                  that is used in the pipeline.
-        :param random_state: The random_state value for the
-                             :py:class:`sklearn.svm.SVC` estimator that is used
-                             in the pipeline.
-        :param scale: How to scale the data before input into the estimator.
-                      If no scaling is to be used set this to None.
-        :return: Model parameters
-        '''
-        return super().get_parameters(word_vectors, tokeniser, lower, C,
-                                      random_state, scale)
-
-    @staticmethod
-    def pipeline():
-        return Pipeline([
-            ('union', FeatureUnion([
-                ('left', Pipeline([
-                    ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('right', Pipeline([
-                    ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('target', Pipeline([
-                    ('contexts', Context('target')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ]))
-            ])),
-            ('scale', MinMaxScaler()),
-            ('svm', LinearSVC())
-        ])
-
-    @staticmethod
-    def _get_word_vector_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the word vectors are set
-        in the pipeline.
-        '''
-
-        return ['union__left__word_vectors__vectors',
-                'union__right__word_vectors__vectors',
-                'union__target__word_vectors__vectors']
-
-    @staticmethod
-    def _get_tokeniser_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the tokenisers are set
-        in the pipeline.
-        '''
-
-        return ['union__left__tokens',
-                'union__right__tokens',
-                'union__target__tokens']
-
-
-class TargetDep(TargetInd):
-    '''
-    Target-dep model from `Vo and Zhang 2015 paper
-    <https://ijcai.org/Proceedings/15/Papers/194.pdf>`_.
-    '''
-
-    @classmethod
-    def get_parameters(cls,
-                       word_vectors: List['bella.word_vectors.WordVectors'],
-                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
-                       lower: bool = True, C: float = 0.01,
-                       random_state: int = 42,
-                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
-        '''
-        Transform the given parameters into a dictonary that is accepted as
-        model parameters
-
-        :param word_vectors: A list of one or more word vectors to be used as
-                             feature vector lookups. If more than one is used
-                             the word vectors are concatenated together to
-                             create a the feature vector for each word.
-        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
-        :param lower: Wether to lower case the words
-        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
-                  that is used in the pipeline.
-        :param random_state: The random_state value for the
-                             :py:class:`sklearn.svm.SVC` estimator that is
-                             used in the pipeline.
-        :param scale: How to scale the data before input into the estimator.
-                      If no scaling is to be used set this to None.
-        :return: Model parameters
-        '''
-        return super().get_parameters(word_vectors, tokeniser, lower, C,
-                                      random_state, scale)
-
-    @staticmethod
-    def pipeline() -> Pipeline:
-        return Pipeline([
-            ('union', FeatureUnion([
-                ('left', Pipeline([
-                    ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('right', Pipeline([
-                    ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('target', Pipeline([
-                    ('contexts', Context('target')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('full', Pipeline([
-                    ('contexts', Context('full')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ]))
-            ])),
-            ('scale', MinMaxScaler()),
-            ('svm', LinearSVC())
-        ])
-
-    @staticmethod
-    def _get_word_vector_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the word vectors are set
-        in the pipeline.
-        '''
-
-        return ['union__left__word_vectors__vectors',
-                'union__right__word_vectors__vectors',
-                'union__target__word_vectors__vectors',
-                'union__full__word_vectors__vectors']
-
-    @staticmethod
-    def _get_tokeniser_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the tokenisers are set
-        in the pipeline.
-        '''
-
-        return ['union__left__tokens',
-                'union__right__tokens',
-                'union__target__tokens',
-                'union__full__tokens']
-
-
-class TargetDepPlus(TargetInd):
-    '''
-    Target-dep+ model from `Vo and Zhang 2015 paper
-    <https://ijcai.org/Proceedings/15/Papers/194.pdf>`_.
-    '''
-
-    @classmethod
-    def get_parameters(cls,
-                       word_vectors: List['bella.word_vectors.WordVectors'],
-                       senti_lexicon: 'bella.lexicons.Lexicon',
-                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
-                       lower: bool = True, C: float = 0.01,
-                       random_state: int = 42,
-                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
-        '''
-        Transform the given parameters into a dictonary that is accepted as
-        model parameters
-
-        :param word_vectors: A list of one or more word vectors to be used as
-                             feature vector lookups. If more than one is used
-                             the word vectors are concatenated together to
-                             create a the feature vector for each word.
-        :param senti_lexicon: Sentiment Lexicon to be used for the Left and
-                              Right sentiment context (LS and RS).
-        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
-        :param lower: Whether to lower case the words
-        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
-                  that is used in the pipeline.
-        :param random_state: The random_state value for the
-                             :py:class:`sklearn.svm.SVC` estimator that is used
-                             in the pipeline.
-        :param scale: How to scale the data before input into the estimator.
-                      If no scaling is to be used set this to None.
-        :return: Model parameters
-        '''
-        params_dict = super().get_parameters(word_vectors, tokeniser, lower, C,
-                                             random_state, scale)
-        params_dict = cls._add_to_params_dict(params_dict,
-                                              cls._get_word_senti_names(),
-                                              senti_lexicon)
-        return params_dict
-
-    @classmethod
-    def get_cv_parameters(cls,
-                          word_vectors: List[List['bella.word_vectors\
-                                                   .WordVectors']],
-                          senti_lexicon: List['bella.lexicons.Lexicon'],
-                          tokeniser=None, lower=None,
-                          C=None, scale=None, random_state=None):
-        '''
-        Transform the given parameters into a list of dictonaries that is
-        accepted as `param_grid` parameter in
-        :py:class:`sklearn.model_selection.GridSearchCV`
-
-        :param word_vectors: A list of a list of word vectors e.g. [[SSWE()],
-                             [SSWE(), GloveCommonCrawl()]]
-        :param senti_lexicon: A list of Sentiment Lexicons to be explored for
-                              the Left and Right sentiment context (LS and RS).
-        :param tokenisers: A list of tokeniser to be used
-                           e.g. :py:meth:`str.split`. Default None, only use
-                           default tokeniser.
-        :param lowers: A list of bool values which indicate whether to lower
-                       case the input words. Default None, only use default
-                       setting(True).
-        :param C: A list of C values for the :py:class:`sklearn.svm.SVC`
-                  estimator that is used in the pipeline. Default None, only
-                  use default setting(C=0.01).
-        :param random_state: A list of random_state values for the
-                             :py:class:`sklearn.svm.SVC` estimator that is used
-                             in the pipeline. Default None, only use default
-                             setting(random_state=42).
-        :param scale: List of scale values. The list can include
-                      :py:class:`sklearn.preprocessing.MinMaxScaler` type of
-                      clases or None if no scaling is to be used. Default None,
-                      only use default setting
-                      (:py:class:`sklearn.preprocessing.MinMaxScaler`).
-        :return: Parameters to explore through cross validation
-        '''
-        params_list = super().get_cv_parameters(word_vectors, tokeniser,
-                                                lower, C, scale, random_state)
-        params_list = cls._add_to_params(params_list, senti_lexicon,
-                                         cls._get_word_senti_names())
-        return params_list
-
-    @staticmethod
-    def pipeline() -> Pipeline:
-        return Pipeline([
-            ('union', FeatureUnion([
-                ('left', Pipeline([
-                    ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('left_s', Pipeline([
-                    ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser()),
-                    ('filter', LexiconFilter()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('right', Pipeline([
-                    ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('right_s', Pipeline([
-                    ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser()),
-                    ('filter', LexiconFilter()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('target', Pipeline([
-                    ('contexts', Context('target')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ])),
-                ('full', Pipeline([
-                    ('contexts', Context('full')),
-                    ('tokens', ContextTokeniser()),
-                    ('word_vectors', ContextWordVectors()),
-                    ('pool_funcs', FeatureUnion([
-                        ('max_pipe', Pipeline([
-                            ('max', NeuralPooling(matrix_max)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('min_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_min)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('avg_pipe', Pipeline([
-                            ('avg', NeuralPooling(matrix_avg)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('prod_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_prod)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ])),
-                        ('std_pipe', Pipeline([
-                            ('min', NeuralPooling(matrix_std)),
-                            ('join', JoinContextVectors(matrix_median))
-                        ]))
-                    ]))
-                ]))
-            ])),
-            ('scale', MinMaxScaler()),
-            ('svm', LinearSVC())
-        ])
-
-    @staticmethod
-    def _get_word_vector_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the word vectors are set
-        in the pipeline.
-        '''
-
-        return ['union__left__word_vectors__vectors',
-                'union__left_s__word_vectors__vectors',
-                'union__right__word_vectors__vectors',
-                'union__right_s__word_vectors__vectors',
-                'union__target__word_vectors__vectors',
-                'union__full__word_vectors__vectors']
-
-    @staticmethod
-    def _get_tokeniser_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the tokenisers are set
-        in the pipeline.
-        '''
-
-        return ['union__left__tokens',
-                'union__left_s__tokens',
-                'union__right__tokens',
-                'union__right_s__tokens',
-                'union__target__tokens',
-                'union__full__tokens']
-
-    @staticmethod
-    def _get_word_senti_names() -> List[str]:
-        '''
-        :returns: A list of of parameter names where the sentiment lexicons
-        are set in the pipeline.
-        '''
-
-        return ['union__left_s__filter__lexicon',
-                'union__right_s__filter__lexicon']
