+++ bella/models/tdparse.py
@@ -1,41 +1,64 @@
 '''
-Contains classes of models that can be found in `Wang et al. paper \
+Module contains all of the classes that represent Machine Learning models
+that are within `Wang et al. paper \
 <https://aclanthology.coli.uni-saarland.de/papers/E17-1046/e17-1046>`_.
 
-Classes:
-
-1. TDParseMinus
-2. TDParse
-3. TDParsePlus
+1. :py:class:`bella.models.target.TDParseMinus` -- TDParse Minus model
+2. :py:class:`bella.models.target.TDParse` -- TDParse model
+3. :py:class:`bella.models.tdparse.TDParsePlus` -- TDParse Plus model
 '''
+from typing import Any, List, Callable, Dict
 
-from sklearn.pipeline import FeatureUnion
-from sklearn.pipeline import Pipeline
-from sklearn.svm import LinearSVC
+from sklearn.pipeline import FeatureUnion, Pipeline
 from sklearn.preprocessing import MinMaxScaler
+from sklearn.svm import LinearSVC
 
-from bella.models.target import TargetInd
-
+import bella
 from bella.tokenisers import ark_twokenize
-from bella.neural_pooling import matrix_max, matrix_min, matrix_avg,\
-matrix_median, matrix_prod, matrix_std
-
+from bella.neural_pooling import matrix_max, matrix_min, matrix_avg
+from bella.neural_pooling import matrix_median, matrix_prod, matrix_std
 from bella.scikit_features.context import Context
 from bella.scikit_features import syntactic_context
+from bella.scikit_features.lexicon_filter import LexiconFilter
+from bella.scikit_features.join_context_vectors import JoinContextVectors
+from bella.scikit_features.neural_pooling import NeuralPooling
 from bella.scikit_features.tokeniser import ContextTokeniser
 from bella.scikit_features.word_vector import ContextWordVectors
-from bella.scikit_features.lexicon_filter import LexiconFilter
-from bella.scikit_features.neural_pooling import NeuralPooling
-from bella.scikit_features.join_context_vectors import JoinContextVectors
+from bella.models.target import TargetInd
+
 
 class TDParseMinus(TargetInd):
-    def __init__(self, child_relations=False):
-        super().__init__()
-        self.child_relations = child_relations
-        self.pipeline = Pipeline([
+
+    def __repr__(self):
+        return 'TDParse Minus'
+
+    def __init__(self, word_vectors: List['bella.word_vectors.WordVectors'],
+                 parser: Any,
+                 tokeniser: Callable[[str], List[str]] = ark_twokenize,
+                 lower: bool = True, C: float = 0.01,
+                 random_state: int = 42,
+                 scale: Any = MinMaxScaler()) -> None:
+        '''
+        :param word_vectors: A list of one or more word vectors to be used as
+                             feature vector lookups. If more than one is used
+                             the word vectors are concatenated together to
+                             create a the feature vector for each word.
+        :param parser: The dependency parser to be used.
+        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
+        :param lower: Whether to lower case the words
+        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
+                  that is used in the pipeline.
+        :param random_state: The random_state value for the
+                             :py:class:`sklearn.svm.SVC` estimator that is
+                             used in the pipeline.
+        :param scale: How to scale the data before input into the estimator.
+                      If no scaling is to be used set this to None.
+        '''
+
+        self.model = Pipeline([
             ('dependency_context', syntactic_context.SyntacticContext()),
             ('contexts', syntactic_context.Context('full')),
-            ('tokens', ContextTokeniser(ark_twokenize, True)),
+            ('tokens', ContextTokeniser()),
             ('word_vectors', ContextWordVectors()),
             ('pool_funcs', FeatureUnion([
                 ('max_pipe', Pipeline([
@@ -60,141 +83,137 @@ class TDParseMinus(TargetInd):
                 ]))
             ])),
             ('scale', MinMaxScaler()),
-            ('svm', LinearSVC(C=0.01))
+            ('svm', LinearSVC())
         ])
-        if child_relations:
-            self.pipeline = Pipeline([
-                ('dependency_context', syntactic_context.DependencyChildContext()),
-                ('tokens', ContextTokeniser(ark_twokenize, True)),
-                ('word_vectors', ContextWordVectors()),
-                ('pool_funcs', FeatureUnion([
-                    ('max_pipe', Pipeline([
-                        ('max', NeuralPooling(matrix_max)),
-                        ('join', JoinContextVectors(matrix_median))
-                    ])),
-                    ('min_pipe', Pipeline([
-                        ('min', NeuralPooling(matrix_min)),
-                        ('join', JoinContextVectors(matrix_median))
-                    ])),
-                    ('avg_pipe', Pipeline([
-                        ('avg', NeuralPooling(matrix_avg)),
-                        ('join', JoinContextVectors(matrix_median))
-                    ])),
-                    ('prod_pipe', Pipeline([
-                        ('min', NeuralPooling(matrix_prod)),
-                        ('join', JoinContextVectors(matrix_median))
-                    ])),
-                    ('std_pipe', Pipeline([
-                        ('min', NeuralPooling(matrix_std)),
-                        ('join', JoinContextVectors(matrix_median))
-                    ]))
-                ])),
-                ('scale', MinMaxScaler()),
-                ('svm', LinearSVC(C=0.01))
-            ])
+        # Inherit from SKLearnModel __init__ method
+        # The parameters here go into the self.get_parameters method
+        super(TargetInd, self).__init__(word_vectors, parser, tokeniser, lower,
+                                        C, random_state, scale)
 
-    @staticmethod
-    def _get_dependency_context():
+    @classmethod
+    def get_parameters(cls,
+                       word_vectors: List['bella.word_vectors.WordVectors'],
+                       parser: Any,
+                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
+                       lower: bool = True, C: float = 0.01,
+                       random_state: int = 42,
+                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
         '''
-        Method to be overidden by subclasses as each pipeline will be different
-        and will have a different parameter name for where the dependency parser
-        related process.
+        Transform the given parameters into a dictonary that is accepted as
+        model parameters
 
-        Dependency related process can be either of the following classes:
-
-        1. syntactic_context.DependencyChildContext
-        2. syntactic_context.SyntacticContext
-
-        :returns: A String of the parameter name where the related dependency \
-        classes are set in the pipeline.
-        :rtype: String
+        :param word_vectors: A list of one or more word vectors to be used as
+                             feature vector lookups. If more than one is used
+                             the word vectors are concatenated together to
+                             create a the feature vector for each word.
+        :param parser: The dependency parser to be used.
+        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
+        :param lower: Whether to lower case the words
+        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
+                  that is used in the pipeline.
+        :param random_state: The random_state value for the
+                             :py:class:`sklearn.svm.SVC` estimator that is
+                             used in the pipeline.
+        :param scale: How to scale the data before input into the estimator.
+                      If no scaling is to be used set this to None.
+        :return: Model parameters
         '''
-
-        return 'dependency_context'
-
-    def get_params(self, word_vector, parser, tokeniser=None, token_lower=None,
-                   parser_lower=None, C=None, random_state=None, scale=True,
-                   rel_depth=None):
-        '''
-        Method extended to include parser, parser_lower and rel_depth
-        parameters.
-        '''
-        params_dict = super().get_params(word_vector, tokeniser=tokeniser,
-                                         lower=token_lower, C=C,
-                                         random_state=random_state, scale=scale)
-        if self.child_relations:
-            if rel_depth is not None:
-                if not isinstance(rel_depth, tuple):
-                    raise TypeError('rel_depth has to be a tuple not {}'\
-                                    .format(type(rel_depth)))
-                rel_depth_name = [self._get_dependency_context() +
-                                  '__rel_depth']
-                params_dict = self._add_to_params_dict(params_dict,
-                                                       rel_depth_name,
-                                                       rel_depth)
-        else:
-            if rel_depth is not None:
-                error_msg = '''rel_depth has been set but will not be used as
-                               you are using the Syntactic Context (full
-                               dependency tree) if you wish to use the dependency
-                               child context in the constructor of this class
-                               set child_relations=True'''
-                ValueError(error_msg)
-        if parser_lower is not None:
-            parser_lower_name = [self._get_dependency_context() + '__lower']
-            params_dict = self._add_to_params_dict(params_dict,
-                                                   parser_lower_name,
-                                                   parser_lower)
-        parser_name = [self._get_dependency_context() + '__parser']
-        params_dict = self._add_to_params_dict(params_dict, parser_name, parser)
-        return params_dict
-
-    def get_cv_params(self, word_vectors, parsers, tokenisers=None,
-                      token_lowers=None, C=None, scale=None, random_state=None,
-                      parser_lowers=None, rel_depths=None):
-        '''
-        Method extended to include parser, parser_lower and rel_depth
-        parameters.
-        '''
-        params_list = super().get_cv_params(word_vectors, tokenisers=tokenisers,
-                                            lowers=token_lowers, C=C, scale=scale,
-                                            random_state=random_state)
-        if self.child_relations:
-            if rel_depths is not None:
-                if not isinstance(rel_depths[0], tuple):
-                    raise TypeError('rel_depths has to be a list of tuples not {}'\
-                                    .format(type(rel_depths[0])))
-                rel_depth_name = self._get_dependency_context() + '__rel_depth'
-                params_list = self._add_to_all_params(params_list,
-                                                      rel_depth_name, rel_depths)
-        else:
-            if rel_depths is not None:
-                error_msg = '''rel_depths has been set but will not be used as
-                               you are using the Syntactic Context (full
-                               dependency tree) if you wish to use the dependency
-                               child context in the constructor of this class
-                               set child_relations=True'''
-                ValueError(error_msg)
-        if parser_lowers is not None:
-            parser_lower_name = self._get_dependency_context() + '__lower'
-            params_list = self._add_to_all_params(params_list, parser_lower_name,
-                                                  parser_lowers)
-        parser_name = self._get_dependency_context() + '__parser'
-        params_list = self._add_to_all_params(params_list, parser_name, parsers)
+        params_list = super().get_parameters(word_vectors, tokeniser,
+                                             lower, C, random_state, scale)
+        params_list = cls._add_to_params_dict(params_list,
+                                              cls._get_dependency_context(),
+                                              parser)
         return params_list
 
-    def __repr__(self):
-        return 'TDParse Minus'
+    @classmethod
+    def get_cv_parameters(cls,
+                          word_vectors: List['bella.word_vectors.WordVectors'],
+                          parser: List[Any],
+                          tokeniser=None, lower=None,
+                          C=None, random_state=None, scale=None):
+        '''
+        Transform the given parameters into a list of dictonaries that is
+        accepted as `param_grid` parameter in
+        :py:class:`sklearn.model_selection.GridSearchCV`
+
+        :param word_vectors: A list of a list of word vectors e.g. [[SSWE()],
+                             [SSWE(), GloveCommonCrawl()]]
+        :param parser: A list of dependency parser to be used.
+        :param tokenisers: A list of tokeniser to be used
+                           e.g. :py:meth:`str.split`. Default None, only use
+                           default tokeniser.
+        :param lowers: A list of bool values which indicate whether to lower
+                       case the input words. Default None, only use default
+                       setting(True).
+        :param C: A list of C values for the :py:class:`sklearn.svm.SVC`
+                  estimator that is used in the pipeline. Default None, only
+                  use default setting(C=0.01).
+        :param random_state: A list of random_state values for the
+                             :py:class:`sklearn.svm.SVC` estimator that is used
+                             in the pipeline. Default None, only use default
+                             setting(random_state=42).
+        :param scale: List of scale values. The list can include
+                      :py:class:`sklearn.preprocessing.MinMaxScaler` type of
+                      clases or None if no scaling is to be used. Default None,
+                      only use default setting
+                      (:py:class:`sklearn.preprocessing.MinMaxScaler`).
+        :return: Parameters to explore through cross validation
+        '''
+        params_list = super().get_cv_parameters(word_vectors, tokeniser,
+                                                lower, C, random_state, scale)
+        params_list = cls._add_to_all_params(params_list,
+                                             cls._get_dependency_context()[0],
+                                             parser)
+        return params_list
+
+    @staticmethod
+    def _get_dependency_context() -> List[str]:
+        '''
+        Method to be overidden by subclasses as each pipeline will be different
+        and will have a different parameter name for where the dependency
+        parser process.
+
+        :returns: A list of parameters names where the dependency parser is
+                  set in the pipeline
+        '''
+
+        return ['dependency_context__parser']
+
 
 class TDParse(TDParseMinus):
-    def __init__(self, child_relations=False):
-        super().__init__(child_relations)
-        self.pipeline = Pipeline([
+
+    def __repr__(self):
+        return 'TDParse'
+
+    def __init__(self, word_vectors: List['bella.word_vectors.WordVectors'],
+                 parser: Any,
+                 tokeniser: Callable[[str], List[str]] = ark_twokenize,
+                 lower: bool = True, C: float = 0.01,
+                 random_state: int = 42,
+                 scale: Any = MinMaxScaler()) -> None:
+        '''
+        :param word_vectors: A list of one or more word vectors to be used as
+                             feature vector lookups. If more than one is used
+                             the word vectors are concatenated together to
+                             create a the feature vector for each word.
+        :param parser: The dependency parser to be used.
+        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
+        :param lower: Whether to lower case the words
+        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
+                  that is used in the pipeline.
+        :param random_state: The random_state value for the
+                             :py:class:`sklearn.svm.SVC` estimator that is
+                             used in the pipeline.
+        :param scale: How to scale the data before input into the estimator.
+                      If no scaling is to be used set this to None.
+        '''
+
+        self.model = Pipeline([
             ('union', FeatureUnion([
                 ('dependency', Pipeline([
                     ('context', syntactic_context.SyntacticContext()),
                     ('contexts', syntactic_context.Context('full')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -221,7 +240,7 @@ class TDParse(TDParseMinus):
                 ])),
                 ('left', Pipeline([
                     ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -248,7 +267,7 @@ class TDParse(TDParseMinus):
                 ])),
                 ('right', Pipeline([
                     ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -275,7 +294,7 @@ class TDParse(TDParseMinus):
                 ])),
                 ('target', Pipeline([
                     ('contexts', Context('target')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -302,146 +321,30 @@ class TDParse(TDParseMinus):
                 ]))
             ])),
             ('scale', MinMaxScaler()),
-            ('svm', LinearSVC(C=0.01))
+            ('svm', LinearSVC())
         ])
-        if child_relations:
-            self.pipeline = Pipeline([
-                ('union', FeatureUnion([
-                    ('dependency', Pipeline([
-                        ('context', syntactic_context.DependencyChildContext()),
-                        ('contexts', syntactic_context.Context('full')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('left', Pipeline([
-                        ('contexts', Context('left')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('right', Pipeline([
-                        ('contexts', Context('right')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('target', Pipeline([
-                        ('contexts', Context('target')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ]))
-                ])),
-                ('scale', MinMaxScaler()),
-                ('svm', LinearSVC(C=0.01))
-            ])
-    @staticmethod
-    def _get_word_vector_names():
-        '''
-        Overideen
+        # Inherit from SKLearnModel __init__ method
+        # The parameters here go into the self.get_parameters method
+        super(TargetInd, self).__init__(word_vectors, parser, tokeniser, lower,
+                                        C, random_state, scale)
 
-        :returns: A list of of parameter names where the word vectors are set in \
-        the pipeline.
-        :rtype: list
+    @staticmethod
+    def _get_word_vector_names() -> List[str]:
+        '''
+        :returns: A list of of parameter names where the word vectors are set
+        in the pipeline.
         '''
 
         return ['union__dependency__word_vectors__vectors',
                 'union__left__word_vectors__vectors',
                 'union__right__word_vectors__vectors',
                 'union__target__word_vectors__vectors']
-    @staticmethod
-    def _get_tokeniser_names():
-        '''
-        Overideen
 
-        :returns: A list of of parameter names where the tokenisers are set in \
-        the pipeline.
-        :rtype: list
+    @staticmethod
+    def _get_tokeniser_names() -> List[str]:
+        '''
+        :returns: A list of of parameter names where the tokenisers are set
+        in the pipeline.
         '''
 
         return ['union__dependency__tokens',
@@ -450,38 +353,51 @@ class TDParse(TDParseMinus):
                 'union__target__tokens']
 
     @staticmethod
-    def _get_dependency_context():
+    def _get_dependency_context() -> List[str]:
         '''
-        Overideen
-
-        Method to be overidden by subclasses as each pipeline will be different
-        and will have a different parameter name for where the dependency parser
-        related process.
-
-        Dependency related process can be either of the following classes:
-
-        1. syntactic_context.DependencyChildContext
-        2. syntactic_context.SyntacticContext
-
-        :returns: A String of the parameter name where the related dependency \
-        classes are set in the pipeline.
-        :rtype: String
+        :returns: A list of parameters names where the dependency parser is
+                  set in the pipeline
         '''
 
-        return 'union__dependency__context'
+        return ['union__dependency__context__parser']
+
+
+class TDParsePlus(TDParseMinus):
 
     def __repr__(self):
-        return 'TDParse'
+        return 'TDParsePlus'
 
-class TDParsePlus(TDParse):
-    def __init__(self, child_relations=False):
-        super().__init__(child_relations)
-        self.pipeline = Pipeline([
+    def __init__(self, word_vectors: List['bella.word_vectors.WordVectors'],
+                 parser: Any, senti_lexicon: 'bella.lexicons.Lexicon',
+                 tokeniser: Callable[[str], List[str]] = ark_twokenize,
+                 lower: bool = True, C: float = 0.01,
+                 random_state: int = 42,
+                 scale: Any = MinMaxScaler()) -> None:
+        '''
+        :param word_vectors: A list of one or more word vectors to be used as
+                             feature vector lookups. If more than one is used
+                             the word vectors are concatenated together to
+                             create a the feature vector for each word.
+        :param parser: The dependency parser to be used.
+        :param senti_lexicon: Sentiment Lexicon to be used for the Left and
+                              Right sentiment context (LS and RS).
+        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
+        :param lower: Whether to lower case the words
+        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
+                  that is used in the pipeline.
+        :param random_state: The random_state value for the
+                             :py:class:`sklearn.svm.SVC` estimator that is
+                             used in the pipeline.
+        :param scale: How to scale the data before input into the estimator.
+                      If no scaling is to be used set this to None.
+        '''
+
+        self.model = Pipeline([
             ('union', FeatureUnion([
                 ('dependency', Pipeline([
                     ('context', syntactic_context.SyntacticContext()),
                     ('contexts', syntactic_context.Context('full')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -508,7 +424,7 @@ class TDParsePlus(TDParse):
                 ])),
                 ('left', Pipeline([
                     ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -535,7 +451,7 @@ class TDParsePlus(TDParse):
                 ])),
                 ('right', Pipeline([
                     ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -562,7 +478,7 @@ class TDParsePlus(TDParse):
                 ])),
                 ('left_s', Pipeline([
                     ('contexts', Context('left')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('filter', LexiconFilter()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
@@ -590,7 +506,7 @@ class TDParsePlus(TDParse):
                 ])),
                 ('right_s', Pipeline([
                     ('contexts', Context('right')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('filter', LexiconFilter()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
@@ -618,7 +534,7 @@ class TDParsePlus(TDParse):
                 ])),
                 ('target', Pipeline([
                     ('contexts', Context('target')),
-                    ('tokens', ContextTokeniser(ark_twokenize, True)),
+                    ('tokens', ContextTokeniser()),
                     ('word_vectors', ContextWordVectors()),
                     ('pool_funcs', FeatureUnion([
                         ('max_pipe', Pipeline([
@@ -645,188 +561,102 @@ class TDParsePlus(TDParse):
                 ]))
             ])),
             ('scale', MinMaxScaler()),
-            ('svm', LinearSVC(C=0.01))
+            ('svm', LinearSVC())
         ])
-        if child_relations:
-            self.pipeline = Pipeline([
-                ('union', FeatureUnion([
-                    ('dependency', Pipeline([
-                        ('context', syntactic_context.DependencyChildContext()),
-                        ('contexts', syntactic_context.Context('full')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('left', Pipeline([
-                        ('contexts', Context('left')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('right', Pipeline([
-                        ('contexts', Context('right')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('left_s', Pipeline([
-                        ('contexts', Context('left')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('filter', LexiconFilter()),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('right_s', Pipeline([
-                        ('contexts', Context('right')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('filter', LexiconFilter()),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ])),
-                    ('target', Pipeline([
-                        ('contexts', Context('target')),
-                        ('tokens', ContextTokeniser(ark_twokenize, True)),
-                        ('word_vectors', ContextWordVectors()),
-                        ('pool_funcs', FeatureUnion([
-                            ('max_pipe', Pipeline([
-                                ('max', NeuralPooling(matrix_max)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('min_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_min)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('avg_pipe', Pipeline([
-                                ('avg', NeuralPooling(matrix_avg)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('prod_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_prod)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ])),
-                            ('std_pipe', Pipeline([
-                                ('min', NeuralPooling(matrix_std)),
-                                ('join', JoinContextVectors(matrix_median))
-                            ]))
-                        ]))
-                    ]))
-                ])),
-                ('scale', MinMaxScaler()),
-                ('svm', LinearSVC(C=0.01))
-            ])
-    @staticmethod
-    def _get_word_vector_names():
-        '''
-        Overideen
+        # Inherit from SKLearnModel __init__ method
+        # The parameters here go into the self.get_parameters method
+        super(TargetInd, self).__init__(word_vectors, parser, senti_lexicon,
+                                        tokeniser, lower, C, random_state,
+                                        scale)
 
-        :returns: A list of of parameter names where the word vectors are set in \
-        the pipeline.
-        :rtype: list
+    @classmethod
+    def get_parameters(cls,
+                       word_vectors: List['bella.word_vectors.WordVectors'],
+                       parser: Any,
+                       senti_lexicon: 'bella.lexicons.Lexicon',
+                       tokeniser: Callable[[str], List[str]] = ark_twokenize,
+                       lower: bool = True, C: float = 0.01,
+                       random_state: int = 42,
+                       scale: Any = MinMaxScaler()) -> Dict[str, Any]:
+        '''
+        Transform the given parameters into a dictonary that is accepted as
+        model parameters
+
+        :param word_vectors: A list of one or more word vectors to be used as
+                             feature vector lookups. If more than one is used
+                             the word vectors are concatenated together to
+                             create a the feature vector for each word.
+        :param parser: The dependency parser to be used.
+        :param senti_lexicon: Sentiment Lexicon to be used for the Left and
+                              Right sentiment context (LS and RS).
+        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`
+        :param lower: Whether to lower case the words
+        :param C: The C value for the :py:class:`sklearn.svm.SVC` estimator
+                  that is used in the pipeline.
+        :param random_state: The random_state value for the
+                             :py:class:`sklearn.svm.SVC` estimator that is used
+                             in the pipeline.
+        :param scale: How to scale the data before input into the estimator.
+                      If no scaling is to be used set this to None.
+        :return: Model parameters
+        '''
+        params_dict = super().get_parameters(word_vectors, parser, tokeniser,
+                                             lower, C, random_state, scale)
+        params_dict = cls._add_to_params_dict(params_dict,
+                                              cls._get_word_senti_names(),
+                                              senti_lexicon)
+        return params_dict
+
+    @classmethod
+    def get_cv_parameters(cls,
+                          word_vectors: List[List['bella.word_vectors\
+                                                   .WordVectors']],
+                          parser: List[Any],
+                          senti_lexicon: List['bella.lexicons.Lexicon'],
+                          tokeniser=None, lower=None,
+                          C=None, random_state=None, scale=None):
+        '''
+        Transform the given parameters into a list of dictonaries that is
+        accepted as `param_grid` parameter in
+        :py:class:`sklearn.model_selection.GridSearchCV`
+
+        :param word_vectors: A list of a list of word vectors e.g. [[SSWE()],
+                             [SSWE(), GloveCommonCrawl()]]
+        :param parser: A list of dependency parser to be used.
+        :param senti_lexicon: A list of Sentiment Lexicons to be explored for
+                              the Left and Right sentiment context (LS and RS).
+        :param tokenisers: A list of tokeniser to be used
+                           e.g. :py:meth:`str.split`. Default None, only use
+                           default tokeniser.
+        :param lowers: A list of bool values which indicate whether to lower
+                       case the input words. Default None, only use default
+                       setting(True).
+        :param C: A list of C values for the :py:class:`sklearn.svm.SVC`
+                  estimator that is used in the pipeline. Default None, only
+                  use default setting(C=0.01).
+        :param random_state: A list of random_state values for the
+                             :py:class:`sklearn.svm.SVC` estimator that is used
+                             in the pipeline. Default None, only use default
+                             setting(random_state=42).
+        :param scale: List of scale values. The list can include
+                      :py:class:`sklearn.preprocessing.MinMaxScaler` type of
+                      clases or None if no scaling is to be used. Default None,
+                      only use default setting
+                      (:py:class:`sklearn.preprocessing.MinMaxScaler`).
+        :return: Parameters to explore through cross validation
+        '''
+        params_list = super().get_cv_parameters(word_vectors, parser,
+                                                tokeniser, lower, C,
+                                                random_state, scale)
+        params_list = cls._add_to_params(params_list, senti_lexicon,
+                                         cls._get_word_senti_names())
+        return params_list
+
+    @staticmethod
+    def _get_word_vector_names() -> List[str]:
+        '''
+        :returns: A list of of parameter names where the word vectors are set
+        in the pipeline.
         '''
 
         return ['union__dependency__word_vectors__vectors',
@@ -835,14 +665,12 @@ class TDParsePlus(TDParse):
                 'union__target__word_vectors__vectors',
                 'union__right_s__word_vectors__vectors',
                 'union__left_s__word_vectors__vectors']
-    @staticmethod
-    def _get_tokeniser_names():
-        '''
-        Overideen
 
-        :returns: A list of of parameter names where the tokenisers are set in \
-        the pipeline.
-        :rtype: list
+    @staticmethod
+    def _get_tokeniser_names() -> List[str]:
+        '''
+        :returns: A list of of parameter names where the tokenisers are set
+        in the pipeline.
         '''
 
         return ['union__dependency__tokens',
@@ -851,48 +679,22 @@ class TDParsePlus(TDParse):
                 'union__target__tokens',
                 'union__right_s__tokens',
                 'union__left_s__tokens']
+
     @staticmethod
-    def _get_word_senti_names():
+    def _get_dependency_context() -> List[str]:
         '''
-        :returns: A list of of parameter names where the sentiment lexicons are \
-        set in the pipeline.
-        :rtype: list
+        :returns: A list of parameters names where the dependency parser is
+                  set in the pipeline
+        '''
+
+        return ['union__dependency__context__parser']
+
+    @staticmethod
+    def _get_word_senti_names() -> List[str]:
+        '''
+        :returns: A list of of parameter names where the sentiment lexicons
+        are set in the pipeline.
         '''
 
         return ['union__left_s__filter__lexicon',
                 'union__right_s__filter__lexicon']
-    def get_params(self, word_vector, parser, senti_lexicon, tokeniser=None,
-                   token_lower=None, parser_lower=None, C=None,
-                   random_state=None, scale=True, rel_depth=None):
-        '''
-        Method extended to include senti_lexicon
-        '''
-        params_dict = super().get_params(word_vector, parser=parser,
-                                         tokeniser=tokeniser, C=C, scale=scale,
-                                         token_lower=token_lower, rel_depth=rel_depth,
-                                         parser_lower=parser_lower,
-                                         random_state=random_state)
-
-        params_dict = self._add_to_params_dict(params_dict,
-                                               self._get_word_senti_names(),
-                                               senti_lexicon)
-        return params_dict
-
-    def get_cv_params(self, word_vectors, parsers, senti_lexicons, tokenisers=None,
-                      token_lowers=None, C=None, scale=None, random_state=None,
-                      parser_lowers=None, rel_depths=None):
-        '''
-        Method extended to include senti_lexicon
-        '''
-        params_list = super().get_cv_params(word_vectors, parsers=parsers,
-                                            tokenisers=tokenisers, C=C, scale=scale,
-                                            token_lowers=token_lowers,
-                                            rel_depths=rel_depths,
-                                            parser_lowers=parser_lowers,
-                                            random_state=random_state)
-        params_list = self._add_to_params(params_list, senti_lexicons,
-                                          self._get_word_senti_names())
-        return params_list
-
-    def __repr__(self):
-        return 'TDParse Plus'
